requirements.txt
alembic==1.7.6
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.0
bleach==4.1.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
distlib==0.3.4
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
dnspython==2.2.1
email-validator==1.1.3
filelock==3.6.0
Flask==2.0.3
Flask-Admin==1.6.0
Flask-Alembic==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
flask-restplus==0.13.0
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-Security-Too==4.1.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
install==1.3.5
itsdangerous==2.1.0
Jinja2==3.0.3
jsonschema==4.4.0
Mako==1.1.6
MarkupSafe==2.1.0
marshmallow==3.14.1
packaging==21.3
passlib==1.7.4
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
screen==1.0.1
six==1.16.0
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
stripe==2.68.0
urllib3==1.26.9
virtualenv==20.13.3
webencodings==0.5.1
Werkzeug==2.0.3
WTForms==3.0.1
xlwt==1.3.0
zipp==3.7.0
